fclCombineList003dca10 MATCH (FUN_0040e3c0_i ABI alias, mask inversion, scoped loop index); mt_evtMisc c540 nd177->128

+++ src/Battle/btlMain.c
@@ -789,7 +789,7 @@ void btlMainInitStateUnitLoad(BtlStateWork* work)
     u64 formationPacketUID;
     u64 lastPacketUID;
     u16 enemyCount;
-    u8 firstUnit;
+    u32 firstUnit;
     enemyCount = 0;
     for (unit = gBtl->unitLists[1].head; unit != ((void*)0); unit = unit->next)
     {
